+++ services/builder/domain/usecases/utils/utils.go
@@ -4,6 +4,7 @@ import (
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb"
 	cbox "github.com/berops/claudie/services/context-box/client"
+	"github.com/rs/zerolog/log"
 )
 
 // SaveConfigWithWorkflowError saves config with workflow states
@@ -41,4 +42,60 @@ func getNodeMap(nodes []*pb.Node) map[string]*pb.Node {
 		m[node.Name] = node
 	}
 	return m
+}
+
+// separateNodepools creates two slices of node names, one for master and one for worker nodes
+func SeparateNodepools(clusterNodes map[string]int32, currentClusterInfo, desiredClusterInfo *pb.ClusterInfo) (master []string, worker []string) {
+	for _, nodepool := range currentClusterInfo.NodePools {
+		var names = make([]string, 0, len(nodepool.Nodes))
+		if np := nodepool.GetDynamicNodePool(); np != nil {
+			if count, ok := clusterNodes[nodepool.Name]; ok && count > 0 {
+				names = getDynamicNodeNames(nodepool, int(count))
+			}
+		} else if np := nodepool.GetStaticNodePool(); np != nil {
+			if count, ok := clusterNodes[nodepool.Name]; ok && count > 0 {
+				names = getStaticNodeNames(nodepool, desiredClusterInfo)
+			}
+		}
+		if nodepool.IsControl {
+			master = append(master, names...)
+		} else {
+			worker = append(worker, names...)
+		}
+	}
+	return master, worker
+}
+
+// getDynamicNodeNames returns slice of length count with names of the nodes from specified nodepool
+// nodes chosen are from the last element in Nodes slice, up to the first one
+func getDynamicNodeNames(np *pb.NodePool, count int) (names []string) {
+	for i := len(np.GetNodes()) - 1; i >= len(np.GetNodes())-count; i-- {
+		names = append(names, np.GetNodes()[i].GetName())
+		log.Debug().Msgf("Choosing node %s for deletion", np.GetNodes()[i].GetName())
+	}
+	return names
+}
+
+// getStaticNodeNames returns slice of length count with names of the nodes from specified nodepool
+// nodes chosen are from the last element in Nodes slice, up to the first one
+func getStaticNodeNames(np *pb.NodePool, desiredCluster *pb.ClusterInfo) (names []string) {
+	// Find desired nodes for node pool.
+	desired := make(map[string]struct{})
+	for _, n := range desiredCluster.NodePools {
+		if n.Name == np.Name {
+			for _, node := range n.Nodes {
+				desired[node.Name] = struct{}{}
+			}
+		}
+	}
+	// Find deleted nodes
+	if n := np.GetStaticNodePool(); n != nil {
+		for _, node := range np.Nodes {
+			if _, ok := desired[node.Name]; !ok {
+				// Append name as it is not defined in desired state.
+				names = append(names, node.Name)
+			}
+		}
+	}
+	return names
 }
